requirements.txt
Django==3.2
python-dotenv
djangorestframework
psycopg2-binary
openai
langchain
psycopg2
langchain_openai
langchain_google_genai
langchain_community
langchain-postgres
django-rest-swagger
drf-yasg
python-dotenv
setuptools
pgvector
ca-vntl-helper
langgraph
django-import-export